Refactored API and updated imports
