+++ packages/block/store/action.js
@@ -1,7 +1,7 @@
-import { ROUND_LENGTH } from '@dpos/validator/constants';
+import { ROUND_LENGTH } from '@dpos/validator/consts';
 import { convertUnixSecondsToLiskEpochSeconds } from '@views/utilities/dateTime';
 import { getBlocks } from '@block/utilities/api';
-import { getForgers, getDelegates } from '@dpos/validator/utils/api';
+import { getForgers, getDelegates } from '@dpos/validator/api';
 import actionTypes from './actionTypes';
 
 /**
