+++ packages/common/utilities/api/search/lsk.js
@@ -3,7 +3,7 @@ import { tokenMap } from '@token/configuration/tokens';
 import { validateAddress } from '@common/utilities/validators';
 import { getAccount } from '@wallet/utilities/api/lsk';
 import { getTransaction } from '@transaction/utilities/api/lsk';
-import { getDelegates } from '@dpos/validator/utils/api/';
+import { getDelegates } from '@dpos/validator/api/';
 import { getBlock } from '@block/utilities/api';
 
 /**
