+++ packages/transaction/utilities/api/lsk.js
@@ -18,7 +18,7 @@ import {
 import { fromRawLsk } from '@token/utilities/lsk';
 import { validateAddress } from '@common/utilities/validators';
 import http from '@common/utilities/api/http';
-import { getDelegates } from '@dpos/validator/utils/api';
+import { getDelegates } from '@dpos/validator/api';
 
 const httpPrefix = '/api/v2';
