+++ packages/transaction/utilities/api/lsk.test.js
@@ -2,7 +2,7 @@
 import { MODULE_ASSETS_NAME_ID_MAP } from '@transaction/configuration/moduleAssets';
 import { getTxAmount } from '@transaction/utilities/transaction';
 import { getState } from '@fixtures/transactions';
-import * as delegates from '@dpos/validator/utils/api';
+import * as delegates from '@dpos/validator/api';
 import http from '@common/utilities/api/http';
 import accounts from '@tests/constants/wallets';
 import {
