+++ packages/views/screens/managers/monitor/delegates/overview/delegatesOverview.js
@@ -1,7 +1,7 @@
 // istanbul ignore file
 import React from 'react';
 // import { fromRawLsk } from '@token/utilities/lsk';
-import { ROUND_LENGTH } from '@dpos/validator/constants';
+import { ROUND_LENGTH } from '@dpos/validator/consts';
 import { useTheme } from '@views/utilities/theme';
 import { getColorPalette } from '@views/basics/charts/chartOptions';
 import Box from '@basics/box';
