+++ packages/views/screens/managers/monitor/delegates/overview/forgingDetails.js
@@ -1,7 +1,7 @@
 // istanbul ignore file
 import React from 'react';
 import moment from 'moment';
-import { ROUND_LENGTH } from '@dpos/validator/constants';
+import { ROUND_LENGTH } from '@dpos/validator/consts';
 import { DoughnutChart } from '@basics/charts';
 import Box from '@basics/box';
 import BoxHeader from '@basics/box/header';
