+++ packages/views/screens/managers/wallet/delegateProfile/performanceView.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavLink } from 'react-router-dom';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 
-import { DEFAULT_STANDBY_THRESHOLD } from '@dpos/validator/constants';
+import { DEFAULT_STANDBY_THRESHOLD } from '@dpos/validator/consts';
 import routes from '@screens/router/routes';
 import { useTheme } from '@views/utilities/theme';
 import { capitalize } from '@common/utilities/helpers';
