+++ src/modules/dpos/validator/consts/index.js
